# contracts/openapi/riverhog.v1.yaml
openapi: 3.1.0
info:
  title: RIVERHOG API
  version: 1.0.0
  summary: MVP API for hot/cold archival storage
  description: Authoritative archive resources are durable across service restarts unless a specific contract explicitly expires or releases them.
servers:
  - url: /v1
paths:
...
  /discs:
    get:
      summary: List burned disc records
      description: Returns one page of generated disc/disc records with sorting and filtering over operator-facing disc metadata.
      operationId: listDiscs
      parameters:
        - in: query
          name: page
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: per_page
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [disc_id, image_id, state, verification_state, location]
            default: id
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - in: query
          name: q
          required: false
          schema: { type: string }
        - in: query
          name: image_id
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Disc list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscListResponse"
        default:
          $ref: "#/components/responses/Error"
  /discs/{disc_id}:
    get:
      summary: Get one burned disc record
      description: Returns one generated disc/disc record by disc id, including disc history.
      operationId: getDisc
      parameters:
        - in: path
          name: disc_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Disc record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscSummary"
        default:
          $ref: "#/components/responses/Error"
  /images:
    get:
      summary: List finalized images
      description: Returns finalized images only, with conventional pagination, sorting, and filtering over finalized-image metadata.
      operationId: listImages
      parameters:
        - in: query
          name: page
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: per_page
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [finalized_at, bytes, discs_registered]
            default: finalized_at
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - in: query
          name: q
          required: false
          schema: { type: string }
        - in: query
          name: collection
          required: false
          schema: { type: string }
        - in: query
          name: has_discs
          required: false
          schema: { type: boolean }
      responses:
        "200":
          description: Finalized image list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalizedImageListResponse"
        default:
          $ref: "#/components/responses/Error"
  /images/{image_id}:
    get:
      summary: Get one image summary
      description: Returns one finalized image summary. `image_id` is the compact UTC finalized id and media identifier.
      operationId: getImage
      parameters:
        - in: path
          name: image_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Image summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalizedImageSummary"
        default:
          $ref: "#/components/responses/Error"
...
  /plan/candidates/{candidate_id}/finalize:
    post:
      summary: Explicitly finalize one provisional candidate
      description: Finalization is the only operation that creates a finalized image id, and the resulting finalized image remains addressable across service restarts.
      operationId: finalizeImage
      parameters:
        - in: path
          name: candidate_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Image finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalizedImageSummary"
        default:
          $ref: "#/components/responses/Error"
  /images/{image_id}/iso:
    get:
      summary: Download ISO bytes for an image
      description: Downloads ISO bytes for an already finalized image. This endpoint is retrieval-only and does not finalize a provisional candidate.
      operationId: getImageIso
      parameters:
        - in: path
          name: image_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: ISO stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        default:
          $ref: "#/components/responses/Error"
  /images/{image_id}/discs:
    get:
      summary: List generated disc slots for an image
      description: Returns the generated disc slots for one finalized image identified by `image_id`.
      operationId: listCopies
      parameters:
        - in: path
          name: image_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Disc listing
          content:
            application/json:
              schema:
                type: object
                required: [discs]
                properties:
                  discs:
                    type: array
                    items:
                      $ref: "#/components/schemas/DiscSummary"
        default:
          $ref: "#/components/responses/Error"
    post:
      summary: Register a physical burned disc
      description: Registers one durable burned-disc record for an already finalized image identified by `image_id`.
      operationId: createCopy
      parameters:
        - in: path
          name: image_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [location]
              properties:
                disc_id:
                  type: string
                  description: Optional generated disc id to claim explicitly for this finalized image.
                location:
                  type: string
                  description: Mutable operator-supplied physical location label. Not part of the disc identity.
      responses:
        "200":
          description: Disc registered
          content:
            application/json:
              schema:
                type: object
                required: [disc]
                properties:
                  disc:
                    $ref: "#/components/schemas/DiscSummary"
        default:
          $ref: "#/components/responses/Error"
  /images/{image_id}/discs/{disc_id}:
    patch:
      summary: Update one generated disc record
      description: Updates mutable operational metadata for one generated disc slot on a finalized image.
      operationId: updateCopy
      parameters:
        - in: path
          name: image_id
          required: true
          schema: { type: string }
        - in: path
          name: disc_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                location:
                  type: string
                  nullable: true
                  description: Updated physical location label. Not part of disc identity.
                state:
                  type: string
                  nullable: true
                  enum: [needed, burning, verified, registered, lost, damaged, retired]
                verification_state:
                  type: string
                  nullable: true
                  enum: [pending, verified, failed]
      responses:
        "200":
          description: Updated disc record
          content:
            application/json:
              schema:
                type: object
                required: [disc]
                properties:
                  disc:
                    $ref: "#/components/schemas/DiscSummary"
        default:
          $ref: "#/components/responses/Error"
  /images/{image_id}/discs/{disc_id}/label-needed:
    post:
      summary: Notify that a disc needs labeling
      description: Sends the best-effort operator notification that a burned and verified disc needs its physical label applied.
      operationId: notifyCopyLabelNeeded
      parameters:
        - in: path
          name: image_id
          required: true
          schema: { type: string }
        - in: path
          name: disc_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Current disc record
          content:
            application/json:
              schema:
                type: object
                required: [disc]
                properties:
                  disc:
                    $ref: "#/components/schemas/DiscSummary"
        default:
          $ref: "#/components/responses/Error"
...
  /jeb/status:
    get:
      summary: Get Jeb collector status
      operationId: getJebStatus
      parameters:
        - in: query
          name: include_backlog
          required: false
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: Jeb status summary
          content:
            application/json:
              schema: { type: object }
        default:
          $ref: "#/components/responses/Error"
  /jeb/attempts:
    get:
      summary: List Jeb attempts
      operationId: listJebBatches
      parameters:
        - in: query
          name: page
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: per_page
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 25 }
        - in: query
          name: sort
          required: false
          schema: { type: string, default: updated_at }
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - in: query
          name: terminal
          required: false
          schema:
            type: string
            enum: [active, terminal, all]
            default: active
        - in: query
          name: state
          required: false
          schema:
            oneOf:
              - type: string
              - type: "null"
        - in: query
          name: account
          required: false
          schema:
            oneOf:
              - type: string
              - type: "null"
        - in: query
          name: collection_slug
          required: false
          schema:
            oneOf:
              - type: string
              - type: "null"
        - in: query
          name: target
          required: false
          schema:
            oneOf:
              - type: string
              - type: "null"
        - in: query
          name: q
          required: false
          schema:
            oneOf:
              - type: string
              - type: "null"
      responses:
        "200":
          description: Paged Jeb attempts
          content:
            application/json:
              schema: { type: object }
        default:
          $ref: "#/components/responses/Error"
  /jeb/config/check:
    get:
      summary: Check deployed Jeb configuration
      operationId: checkJebConfig
      responses:
        "200":
          description: Jeb configuration check result
          content:
            application/json:
              schema: { type: object }
        default:
          $ref: "#/components/responses/Error"
  /jeb/once:
    post:
      summary: Request one Jeb scheduler pass
      operationId: runJebOnce
      responses:
        "202":
          description: Scheduler pass accepted and started by the Jeb service
          content:
            application/json:
              schema: { type: object }
        default:
          $ref: "#/components/responses/Error"
  /jeb/archive-now:
    post:
      summary: Archive one Jeb account immediately
      operationId: archiveJebNow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JebArchiveNowRequest"
      responses:
        "200":
          description: Dry-run archive plan returned without mutating Jeb state
          content:
            application/json:
              schema: { type: object }
        "202":
          description: Archive batch accepted; processing continues on the Jeb service
          content:
            application/json:
              schema: { type: object }
        default:
          $ref: "#/components/responses/Error"
components:
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    JebArchiveNowRequest:
      type: object
      required: [account]
      properties:
        account:
          type: string
          minLength: 1
        process:
          type: boolean
          default: true
        dry_run:
          type: boolean
          default: false
    CoverageState:
      type: string
      enum: [none, partial, full]
...
    FinalizedImageSummary:
      type: object
      required:
        [
          id,
          filename,
          finalized_at,
          bytes,
          target_bytes,
          fill,
          files,
          collections,
          collection_ids,
          iso_ready,
          disc_redundancy_state,
          discs_required,
          discs_registered,
          discs_verified,
          discs_missing,
        ]
      properties:
        id:
          type: string
          pattern: "^[0-9]{8}T[0-9]{6}Z$"
          description: Canonical finalized image id and immutable media identifier.
        filename:
          type: string
          description: Stable ISO filename for the finalized image artifact.
        finalized_at:
          type: string
          format: date-time
          description: Expanded UTC timestamp encoded by the finalized image id.
        bytes: { type: integer }
        target_bytes:
          type: integer
          description: Target media capacity used when this image was finalized.
        fill: { type: number }
        files: { type: integer }
        collections:
          type: integer
          description: Count of collection ids represented by the finalized image.
        collection_ids:
          type: array
          items: { type: string }
          description: Lexically sorted contained collection ids.
        iso_ready:
          type: boolean
          const: true
          description: Finalized images are always ISO-ready.
        disc_redundancy_state:
          $ref: "#/components/schemas/CoverageState"
        discs_required:
          type: integer
          minimum: 1
        discs_registered:
          type: integer
          minimum: 0
        discs_verified:
          type: integer
          minimum: 0
        discs_missing:
          type: integer
          minimum: 0
    FinalizedImageListResponse:
      type: object
      required: [page, per_page, total, pages, sort, order, images]
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
        sort:
          type: string
          enum: [finalized_at, bytes, discs_registered]
        order:
          type: string
          enum: [asc, desc]
        images:
          type: array
          items:
            $ref: "#/components/schemas/FinalizedImageSummary"
...
    DiscSummary:
      type: object
      required:
        [
          disc_id,
          label_text,
          location,
          created_at,
          state,
          verification_state,
          history,
          image_id,
        ]
      properties:
        disc_id:
          type: string
          description: Generated disc id, unique within one finalized image.
        label_text:
          type: string
          description: Exact operator-facing disc label text for this generated disc slot.
        location:
          type: string
          nullable: true
          description: Mutable physical location label. Not part of disc identity.
        created_at: { type: string, format: date-time }
        state:
          type: string
          enum: [needed, burning, verified, registered, lost, damaged, retired]
        verification_state:
          type: string
          enum: [pending, verified, failed]
        history:
          type: array
          items:
            $ref: "#/components/schemas/DiscHistory"
        image_id:
          type: string
          pattern: "^[0-9]{8}T[0-9]{6}Z$"
          description: Finalized image id for this disc.
        filename:
          type: string
          nullable: true
          description: Stable ISO filename for the finalized image artifact.
    DiscListResponse:
      type: object
      required: [page, per_page, total, pages, sort, order, query, discs]
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
        sort:
          type: string
          enum: [disc_id, image_id, state, verification_state, location]
        order:
          type: string
          enum: [asc, desc]
        query:
          type: string
          nullable: true
        image_id:
          type: string
          nullable: true
        discs:
          type: array
          items:
            $ref: "#/components/schemas/DiscSummary"
    DiscHistory:
      type: object
      required: [at, event, state, verification_state, location]
      properties:
        at: { type: string, format: date-time }
        event: { type: string }
        state:
          type: string
          enum: [needed, burning, verified, registered, lost, damaged, retired]
        verification_state:
          type: string
          enum: [pending, verified, failed]
        location:
          type: string
          nullable: true
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }